Extract use case class for creating reminders for appointments
